Extra Game Scene
Finished a scene that will play a simple 'click the shapes' game.

How it works: Tap the screen to start. Tap shapes to add to your score. Try to get the highest score before the time runs out.

Next I am going to include animations and the other optional challenges to the game.